requirements.txt
numpy
Pillow
requests